requirements.txt
fastapi==0.112.2
uvicorn==0.30.6
sqlalchemy==2.0.33
databases==0.9.0
asyncpg==0.29.0
pydantic==2.8.2
starlette==0.38.4
psycopg2-binary==2.9.9
passlib==1.7.4
pyjwt==2.8.0